+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 - [8318](https://github.com/vegaprotocol/vega/issues/8318) - Proto definitions for spots
 - [8117](https://github.com/vegaprotocol/vega/issues/8117) - Added spots governance implementation
 - [8259](https://github.com/vegaprotocol/vega/issues/8259) - Proto definitions for successor markets.
+- [8201](https://github.com/vegaprotocol/vega/issues/8201) - Add support for successor markets.
 
 ### 🐛 Fixes
 - [8208](https://github.com/vegaprotocol/vega/issues/8208) - Fix block explorer API documentation